src/requirements.txt
pandas
numpy
networkx
bokeh.models 
bokeh.plotting